Miscellaneous refactors and changes
- Move changes from Assembly wrapper to new wrapper
- Revert Assembly wrapper changes
- Add "endless wrapper" field for Target Environments
- Create Assembly (with linker script) environment for RISC-V
- Remove "language" option in mp_mpt2test, use environments instead
