+++ targets/generic/tools/mp_mpt2test.py
@@ -55,7 +55,7 @@ LOG = get_logger(__name__)
 
 
 # Functions
-def generate(test_definition, outputfile, ldscriptfile, language, target,
+def generate(test_definition, outputfile, ldscriptfile, target,
              **kwargs):
     """
     Microbenchmark generation policy
@@ -178,18 +178,12 @@ def generate(test_definition, outputfile, ldscriptfile, language, target,
 
     wrapper_kwargs = {}
 
-    if language == "c":
-        if kwargs.get("wrap_function", False):
-            raise MicroprobeException("The wrap-function option "
-                                      "is incompatible with the C langauge "
-                                      "backend.")
+    if kwargs.get("endless", False):
+        wrapper_name = target.environment.default_endless_wrapper
+    else:
+        wrapper_name = target.environment.default_wrapper
 
-        wrapper_name = "C"
-        if kwargs.get("endless", False):
-            print_warning("Using endless c wrapper")
-            wrapper_name = "CInfGen"
-    elif language == "asm":
-        wrapper_name = "Assembly"
+    if wrapper_name == "AssemblyLd":
         if sequence[0].label is not None:
             start_label = sequence[0].label
         else:
@@ -418,13 +412,6 @@ def main():
         group=groupname
     )
 
-    cmdline.add_option(
-        "language",
-        "l",
-        "c",
-        "Test language"
-    )
-
     cmdline.add_flag(
         "safe-bin", None, "Ignore unrecognized binary codifications (do not"
                           "fail). Useful when MPTs are generated by dumping "
@@ -506,7 +493,6 @@ def _main(arguments):
     target = import_definition(arguments.pop('target'))
     test_definition = arguments.pop('mpt_definition')
     outputfile = arguments.pop('output_file')
-    language = arguments.pop('language')
 
     if "output_ld_script" in arguments:
         ldscriptfile = arguments.pop('output_ld_script')
@@ -514,8 +500,7 @@ def _main(arguments):
         ldscriptfile = outputfile + ".lds"
 
     print_info("Start generating '%s'" % outputfile)
-    generate(test_definition, outputfile, ldscriptfile, language, target,
-             **arguments)
+    generate(test_definition, outputfile, ldscriptfile, target, **arguments)
     print_info("'%s' generated!" % outputfile)
